+++ lib/ansible/plugins/action/iosxr.py
@@ -23,13 +23,13 @@ import os
 import sys
 import copy
 
-from ansible.plugins.action.normal import ActionModule as _ActionModule
-from ansible.module_utils._text import to_bytes
-from ansible.utils.path import unfrackpath
-from ansible.plugins import connection_loader
-from ansible.compat.six import iteritems
-from ansible.module_utils.iosxr import iosxr_argument_spec
 from ansible.module_utils.basic import AnsibleFallbackNotFound
+from ansible.module_utils.iosxr import iosxr_argument_spec
+from ansible.module_utils.six import iteritems
+from ansible.module_utils._text import to_bytes
+from ansible.plugins import connection_loader
+from ansible.plugins.action.normal import ActionModule as _ActionModule
+from ansible.utils.path import unfrackpath
 
 try:
     from __main__ import display