+++ lib/ansible/plugins/action/nxos.py
@@ -26,9 +26,9 @@ import copy
 from ansible.plugins.action.normal import ActionModule as _ActionModule
 from ansible.utils.path import unfrackpath
 from ansible.plugins import connection_loader
-from ansible.compat.six import iteritems
-from ansible.module_utils.nxos import nxos_argument_spec
 from ansible.module_utils.basic import AnsibleFallbackNotFound
+from ansible.module_utils.nxos import nxos_argument_spec
+from ansible.module_utils.six import iteritems
 from ansible.module_utils._text import to_bytes
 
 try:
@@ -36,6 +36,7 @@ try:
 except ImportError:
     from ansible.utils.display import Display
     display = Display()
+
 
 class ActionModule(_ActionModule):
 
@@ -56,6 +57,7 @@ class ActionModule(_ActionModule):
             pc = copy.deepcopy(self._play_context)
             pc.connection = 'network_cli'
             pc.network_os = 'nxos'
+            pc.remote_addr = provider['host'] or self._play_context.remote_addr
             pc.port = provider['port'] or self._play_context.port or 22
             pc.remote_user = provider['username'] or self._play_context.connection_user
             pc.password = provider['password'] or self._play_context.password
@@ -114,7 +116,10 @@ class ActionModule(_ActionModule):
         # make sure a transport value is set in args
         self._task.args['transport'] = transport
 
-        return super(ActionModule, self).run(tmp, task_vars)
+        result = super(ActionModule, self).run(tmp, task_vars)
+        del result['invocation']['module_args']['provider']
+
+        return result
 
     def _get_socket_path(self, play_context):
         ssh = connection_loader.get('ssh', class_only=True)